+++ src/anomalib/pipelines/orchestrators/__init__.py
@@ -1,0 +1,9 @@
+"""Orchestrators."""
+
+# Copyright (C) 2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from .base import Orchestrator
+from .benchmark import Benchmark
+
+__all__ = ["Orchestrator", "Benchmark"]
