+++ tools/runners/sample.yaml
@@ -1,17 +1,17 @@
 # sample script to show grid search for two categories
-n_jobs: 2
-seed:
-  grid: [42, 43]
-hardware: cuda
-model:
-  class_path:
-    grid: [Padim, Patchcore]
-
-data:
-  class_path: MVTec
-  init_args:
-    category:
-      grid:
-        - bottle
-        - capsule
-    image_size: [256, 256]
+accelerator:
+  - cuda
+  - cpu
+benchmark:
+  seed: 42
+  model:
+    class_path:
+      grid: [Padim, Patchcore]
+  data:
+    class_path: MVTec
+    init_args:
+      category:
+        grid:
+          - bottle
+          - capsule
+      image_size: [256, 256]
